+++ @utahdts/utah-design-system-header/src/js/renderables/mobile/hookupUtahIdInMobileMenu.js
@@ -59,12 +59,20 @@ export function hookupUtahIdInMobileMenu(mobileMenuWrapper, utahIdPopup) {
         const currentUtahIdData = getCurrentUtahIdData();
         const settings = getUtahHeaderSettings();
         let result = false;
-        if (!currentUtahIdData?.isDefinitive || !currentUtahIdData?.userInfo?.authenticated) {
+        const onSignIn = (settings.utahId !== false && settings.utahId !== true && settings.utahId?.onSignIn);
+
+        if (
+          // check if controlled by app and not logged in
+          (settings.utahId !== false && settings.utahId !== true && !settings.utahId?.currentUser?.authenticated)
+          // check if not controlled by app and not logged in (definitive is true while ajax request in flight)
+          || (settings.utahId === true && (!currentUtahIdData?.isDefinitive || !currentUtahIdData?.userInfo?.authenticated))
+        ) {
           result = true;
-          const onSignIn = (settings.utahId !== false && settings.utahId !== true && settings.utahId?.onSignIn);
           if (onSignIn) {
+            // custom onSignIn passed in settings
             onSignIn(e);
           } else {
+            // generic default sign method
             e.preventDefault();
             e.stopPropagation();
             window.location.href = utahIdUrls.SIGN_IN;