+++ @utahdts/utah-design-system-header/src/js/renderables/popupMenu/renderPopupMenu.js
@@ -20,6 +20,7 @@ import { popupFocusHandler } from '../../misc/popupFocusHandler';
 import { renderDOMSingle } from '../../misc/renderDOMSingle';
 import { uuidv4 } from '../../misc/uuidv4';
 import { renderPopup } from '../popup/renderPopup';
+import { suffixForMenuItemTitle } from '../mainMenu/suffixForMenuItemTitle';
 
 /**
  * @typedef {import('src/@types/jsDocTypes.d').MenuItem} MenuItem
@@ -125,11 +126,8 @@ function renderPopupMenuItem(menuUl, popupMenuItem, options) {
     throw new Error('renderPopupMenuItem: titleSpanLink not found');
   }
 
-  menuAHref.onclick = onClickBlur;
-  menuButton.onclick = onClickBlur;
-
   const actionMenu = popupMenuItem.actionMenu && [...popupMenuItem.actionMenu];
-  // add "(page)" menu item if menu item is an actionMenu && a link
+  // add `parentMenuLinkSuffix` menu item if menu item is an actionMenu && a link
   if (actionMenu && (
     popupMenuItem.actionFunction
     || popupMenuItem.actionUrl
@@ -141,8 +139,13 @@ function renderPopupMenuItem(menuUl, popupMenuItem, options) {
       actionUrl: popupMenuItem.actionUrl,
       className: popupMenuItem.className,
       icon: popupMenuItem.icon,
-      title: `${popupMenuItem.title} (page)`,
+      title: `${popupMenuItem.title}${suffixForMenuItemTitle(popupMenuItem, options.parentMenuLinkSuffix)}`,
     });
+  }
+
+  if (!actionMenu?.length) {
+    menuAHref.onclick = onClickBlur;
+    menuButton.onclick = onClickBlur;
   }
 
   // three types of action: parent, custom function, link