+++ @utahdts/utah-design-system/@types/jsDocTypes.d.js
@@ -52,6 +52,8 @@
 
 /**
  * IconFunc
+ * if getting an error that props is mandatory, use Component syntax instead of functional:
+ * ie <Icons.IconEnvelope /> instead of {Icons.IconEnvelope()}
  * @typedef {(props: {className?: string, altText?: string}) => React.ReactNode} IconFunc
  */
 
@@ -357,7 +359,7 @@
 /**
  * @typedef MultiSelectContextNonStateRef {
  *  @property {HTMLElement | null} comboBoxDivElement the text input of the multi select (for forcing focused)
- *  @property {(HTMLDivElement | null)[]} selectedOptionTagRefs refs to the selected tags elements
+ *  @property {(HTMLLIElement | null)[]} selectedOptionTagRefs refs to the selected tags elements
  *  @property {HTMLInputElement | null} textInput the textInput used for this multi select's combo box
  * }
  */