+++ utah-design-system-website/src/react/components/routing/Routing.jsx
@@ -77,6 +77,7 @@ export function Routing() {
                   <VerticalMenu
                     className="menu-side-panel"
                     currentMenuItem={currentMenuItem}
+                    // @ts-ignore
                     menus={menuSecondary ?? null}
                   />
                 )}